Work on NixOS compilation; instructions to README.
Reason being that it would be nice to have everything compiling without needing a Distrobox container. I'd still need to sort out compiling inside of CLion or another IDE natively on NixOS (I'm thinking on a nix-develop flake, but I still need to experiment around with that), but I want to try and tackle actually getting the mod to compile through NixOS first.

Right now, the default.nix needs some work, for some reason it fails compiling the necessary components to cross-compile, but it seems like it at least gets through the nix-build process somewhat.